+++ lib/xml_templates.dart
@@ -2,15 +2,15 @@ const String icLauncherXml = '''
 <?xml version="1.0" encoding="utf-8"?>
 <adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
   <background android:drawable="@color/ic_launcher_background"/>
-  <foreground android:drawable="@drawable/ic_launcher_foreground"/>
+  <foreground android:drawable="@mipmap/ic_launcher_foreground"/>
 </adaptive-icon>
 ''';
 
-const String icLauncherDrawableBackgroundXml = '''
+const String icLauncherRoundXml = '''
 <?xml version="1.0" encoding="utf-8"?>
 <adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
-  <background android:drawable="@drawable/ic_launcher_background"/>
-  <foreground android:drawable="@drawable/ic_launcher_foreground"/>
+  <background android:drawable="@color/ic_launcher_background"/>
+  <foreground android:drawable="@mipmap/ic_launcher_foreground"/>
 </adaptive-icon>
 ''';